pyproject.toml
[tool.poetry]
name = "groovis"
version = "0.1.0"
description = ""
authors = ["Kwang-myung Yu <sguys99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pillow = "^10.0.0"
torch = "1.11.0"
einops = "^0.6.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.3.0"
pytest = "^7.4.0"
numpy = "^1.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
